Refactor Omnisharp configuration generation to separate project generator
